Rename Movement to MoveCommand and Rotate to RotateCommand
